+++ quick-start/src/main/ui/app/components/flows-new/manage-flows.module.ts
@@ -1,8 +1,8 @@
 import {NgModule} from '@angular/core';
-import {RouterModule, Routes} from '@angular/router';
-import {FlowsPageComponent} from "./manage-flows/flows-page.component";
+import {RouterModule} from '@angular/router';
+import {ManageFlowsComponent} from "./manage-flows/manage-flows.component";
 import {MaterialModule} from "../theme/material.module";
-import {FlowsPageUiComponent} from "./manage-flows/ui/flows-page-ui.component";
+import {ManageFlowsUiComponent} from "./manage-flows/ui/manage-flows-ui.component";
 import {ConfirmationDialogComponent} from "../common";
 import {NewFlowDialogComponent} from "./manage-flows/ui/new-flow-dialog.component";
 import {HttpClientModule} from "@angular/common/http";
@@ -11,8 +11,8 @@ import {HttpClientModule} from "@angular/common/http";
   declarations: [
     ConfirmationDialogComponent,
     NewFlowDialogComponent,
-    FlowsPageUiComponent,
-    FlowsPageComponent
+    ManageFlowsUiComponent,
+    ManageFlowsComponent
   ],
   imports     : [
     MaterialModule,
