+++ quick-start/src/main/ui/app/components/flows-new/models/step.model.ts
