+++ quick-start/src/main/ui/app/components/flows-new/services/manage-flows.service.ts
@@ -1,6 +1,5 @@
 import {Injectable} from '@angular/core';
 import {HttpClient} from '@angular/common/http';
-import {map} from 'rxjs/operators';
 
 
 @Injectable()
